Debian: Drop versioned dep on lsb-base so that the I2P package can be installed in GnewSense.
This change will break nothing since we check for the existence of the lsb
function files in the initscript. With an older version of lsb-base (like the
one that came with Hardy (and GnewSense uses), the output from the initscript
will simply look a little different (but again, nothing will break). Virtually 
everything but GnewSense comes with the desired lsb-base package anyway...